refactor: move `middleware-json` into `middlewares`
