+++ __tests__/middlewares/json/basic.spec.ts
@@ -1,6 +1,6 @@
-import { fatcher } from 'fatcher';
+import { fatcher } from '../../../src';
 import fetchMock from 'jest-fetch-mock';
-import { json } from '../src';
+import { json } from '../../../src/middlewares/json';
 
 describe('Json', () => {
     const BASE_URL = 'https://virual.com';
